remove npcs from entities folder and interface. they now implement INPC
