+++ cse3902/Interfaces/INPC.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using Microsoft.Xna.Framework;
+
+namespace cse3902.Interfaces
+{
+    public interface INPC
+    {
+        public void Update(GameTime gameTime);
+        public void Draw();
+        public void SetMessage(string message);
+    }
+}
