+++ cse3902/cse3902.csproj
@@ -43,9 +43,11 @@
     <Folder Include="Collision\Collidables\" />
     <Folder Include="XMLParsing\" />
     <Folder Include="Entities\DamageMasks\" />
+    <Folder Include="NPCs\" />
   </ItemGroup>
   <ItemGroup>
     <Compile Remove="Blocks\WaterBlock.cs" />
     <Compile Remove="Entities\EnemyStateMachine.cs" />
+    <Compile Remove="Collision\Collidables\NPCCollidable.cs" />
   </ItemGroup>
 </Project>
